Move all datasources to subpackage import level
