+++ zenml/core/datasources/postgres_datasource.py
@@ -15,8 +15,8 @@
 
 from typing import Text, Dict
 
-from zenml.core.datasources.base_datasource import BaseDatasource
-from zenml.core.steps.data.postgres_data_step import PostgresDataStep
+from zenml.core.datasources import BaseDatasource
+from zenml.core.steps.data import PostgresDataStep
 
 
 class PostgresDatasource(BaseDatasource):
